remove EDITOR and use VISUAL insted
VISUAL stands for visual editor.
EDITOR stands for editor.
It's a tad bit misleading due to its age. This env var standard dates
back to when it was not unexpected to have a terminal session in a
session that did not support a editor session, with screen redraws and
special escape codes to draw in regions, and a known screensize, e.g. a
teletypewriter (The thing you press keys on and it typed on paper, and
it would type out the computers response to a command. Imagine if that
tried rewriting all your screen changes over the old character, yikes!).
So back then EDITOR literally meant a simple CLI (not TUI) editor, like
ed. And for the more feature-full sessions, like these new fancy
terminal emulators with their fancy redering, there was the VISUAL var,
for editing a file/buffer with visual features supported.
By default, if a program is compliant with the standard, whatever
program will first try VISUAL, and if this fails (usually due to not
being in a visual session), then it will fall back onto EDITOR.
If you have EDITOR set to a visual editor, then you will have no
fallback causing what in rare situations could've been using a fallback
editor, to instead have total failure.
So I would set VISUAL to nvim, and EDITOR to a non-visual editor, and if
you don't know one at all, then don't set EDITOR or learn how to use a
non-visual editor.